# Day10.  Still took 1.2 seconds. Probably some great optimizations to be had