--- a/rust/src/lib.rs
+++ b/rust/src/lib.rs
@@ -1,6 +1,7 @@
 use aoc_runner_derive::aoc_lib;
 
 mod day1;
+mod day10;
 mod day2;
 mod day3;
 mod day4;