```
AOC 2015
Day 10 - Part 1 : 360154
        generator: 149ns,
        runner: 90.794723ms

Day 10 - Part 2 : 5103798
        generator: 98ns,
        runner: 1.254376903s
```